feat: folders & sets updates
