+++ client/src/modules/folder/FolderDetails.tsx
@@ -7,10 +7,19 @@ import { ActionList } from '@src/ui/ActionList';
 import { MoreIcon } from '@src/ui/Icons';
 import { DropdownMenu } from '@src/ui';
 import { FolderInterface } from '@src/interfaces';
+import { setApi } from '@src/apis';
+import { useQuery } from 'react-query';
 
-const FolderDetails: NextPage<{ queryId: string; data: FolderInterface }> = ({ queryId, data: folder }) => {
+const FolderDetails: NextPage<{ queryId: string; queryUser: string; data: FolderInterface }> = ({
+  queryId,
+  queryUser,
+  data: folder,
+}) => {
   const router = useRouter();
   const [shownModal, setShownModal] = React.useState<'edit' | 'del' | null>(null);
+  React.useEffect(() => {
+    if (queryId === 'new') setShownModal('edit');
+  }, [queryId]);
 
   const menuOptions = [
     { title: 'Author', action: () => router.push(`/user/${folder?.user.id}`) },
@@ -18,9 +27,7 @@ const FolderDetails: NextPage<{ queryId: string; data: FolderInterface }> = ({ q
     { title: 'Delete', action: () => setShownModal('del') },
   ];
 
-  React.useEffect(() => {
-    if (queryId === 'new') setShownModal('edit');
-  }, [queryId]);
+  const userSets = useQuery('userSets', () => setApi.getByUser(String(queryUser)), { enabled: Boolean(queryUser) });
 
   return (
     <>
@@ -46,22 +53,23 @@ const FolderDetails: NextPage<{ queryId: string; data: FolderInterface }> = ({ q
             </div>
           </div>
 
-          <div className="p-4 max-w-3xl mx-auto">
-            <h2 className="text-xl mb-4 font-medium">Sets</h2>
-            <ActionList
-              placeholder={'Nothing yet'}
-              data={folder.sets}
-              keyExtractor={(item) => item.id}
-              renderItem={(item, index) => (
-                <ActionList.Link href={`/sets/${item.id}`} isFirst={index === 0}>
-                  {item.title}
-                </ActionList.Link>
-              )}
-            />
-          </div>
           <FolderDelete data={folder} open={shownModal === 'del'} setOpen={() => setShownModal(null)} />
         </>
       ) : null}
+
+      <div className="p-4 max-w-3xl mx-auto">
+        <h2 className="text-xl mb-4 font-medium">Sets</h2>
+        <ActionList
+          placeholder={'Nothing yet'}
+          data={folder?.sets || userSets.data || []}
+          keyExtractor={(item) => item.id}
+          renderItem={(item, index) => (
+            <ActionList.Link href={`/sets/${item.id}`} isFirst={index === 0}>
+              {item.title}
+            </ActionList.Link>
+          )}
+        />
+      </div>
 
       <FolderForm
         data={folder}
