--- client/src/modules/index.ts
@@ -1,1 +1,0 @@
-export { default as Dashboard } from './Dashboard';
